Disambiguate build directories.
We suffix all the build directories and files that we make with a
package name and version. As package name is a target name, there can
only be a unique one in single project and we preserve some sanity in
build environment, at least for the directories we declare and pass
around. This allows multiple `haskell_library` and `haskell_binary`
targets in a single package.

Closes #7.