Configurable tooling
This allows us to provide tools to bazel rules to build with. By
default we use whatever `@ghc` and `@binutils` are available in the
workspace.

I'm unsure how to document/deal with this properly: right now it may
be surprising to the user to have to populate those. They are no
longer accepted from environment however which I suspect is what we
want to begin with.

Ideally we would provide some `compiler.bzl` with various GHC versions
out of rules_nixpkgs/official bindists. I'll have to look how to do
that and make it available to user just like we can say `@openjdk`
today without much extra work.