Reorganize tests
Use `rule_test` to verify that a rule a) worked, b) produced the
expected output. This obviates the need for a shell script that
launches binaries and is more general, so can be used systematically.
It's the testing strategy that `rules_rust` uses. We may want to
revive launching binaries later, but this is good enough for now.
Also, systematically put each test in its own subdir.