Introduce haskell_import. Replaces external_libraries field.
Prebuilt dynamic libraries, i.e. what Cabal calls `extra-libraries`,
are dependencies like any other. It's just that it's not a Haskell
dependency. The idiomatic way to express this in Bazel is to mix
dependencies of different types all in the `deps` field. So we move to
the content of `external_libraries` to `deps`. Further, for forward
compatibility, we introduce `haskell_import` to mirror `cc_import` and
`java_import`. It doesn't do much at the moment beyond reexport the
shared library it was given and do some sanity checking. But it may
need to do in the future.

This PR introduces two new tests for this functionality. They depend
on `rules_nixpkgs`, which in turn depends on Nix. Since repositories
are loaded lazily, it doesn't mean that Nix is required just to use
the rules, only to test them. We could perhaps conditionally disable
Nix in the future (using constraints?), but this is good enough for
now. And better than the alternative: hard coding absolute paths to
libraries in tests.