Allow hsc files in srcs, remove `hscs` rule attribute.
We can uniquely identify these through an extension anyway.
Technically one could have an arbitrary extension and still want to
treat the file as `hsc` file but we weren't allowing that anyway is
`hscs` field already. This way is friendlier to the user and none of
the code really cared about the distinction anyway.

Fixes #26.