Factor out common build elements of lib/bin components.
This lets us share default build logic between
haskell_library/haskell_binary: this is desirable as we now can't
forget to process some files in one or the other or apply flags in one
or the other &c. Both still do have their own behaviour but it's a
much smaller surface to maintain.

Fixes #20, fixes #13.