Introduce interop with Java rules.
This allows us to list Java dependencies in `deps` of Haskell rules
and have appropriate inputs and CLASSPATH be available during compile
time if for example `javac` is called during build.

Relates to #80 but does not close it: we ideally want transitive
classpath (I think? cc @mboes , maybe we're OK with how things are
now) but accessing the necessary attribute crashes bazel in some
cases. I'll be creating an upstream ticket shortly.