Mangle names of solibs before linking
Hackage libraries often include small amounts of C code. A common
pattern is to package this code as a library, called `cbits` by
convention. Unfortunately, when multiple libraries each have
a `cbits`, then names can clash during linking.

So we introduce a symlink that has a name that is guaranteed to be
globally unique within the workspace. Since the label of the target
that produced a library is globally unique, we use that as
a qualifier. We haven't invented our solution to a common problem.
Indeed because the problem _is_ common, it turns out that built-in CC
rules implement exactly this name mangling, also using symlinks.
Unfortunately, this mangling is not exposed to Skylark rules (see
bazelbuild/bazel#4581). So all we're doing here is emulate the
mangling that the built-in rules do.

Fixes #103.

Closes #116.