Introduce textual_hdrs
We already had this, in the form of `external_deps`, which
5d1a179c0463c182766af1603b38ad04dadef4fb removed. As noted there, we
removed it for lack of a good use case. We just found a use case:
compile aeson without making modifications to the code organization of
that package. The new name for this is consistent with
https://docs.bazel.build/versions/master/be/c-cpp.html#cc_library.textual_hdrs.

Tested with bazel-protolude aeson.