Fix grep not found error
At least in sandbox mode, the ghc-defs-cleanup.sh script fails with
a `grep: command not found` error, because `PATH` is sanitized by
Bazel. We set `use_default_shell` to reuse the host shell environment.
This is an impurity, because it makes the build result depend on
what's on the host's PATH. But we already borrow up `/bin/sh` from the
host. Arguably it's also reasonable to borrow `grep`.