Resolve the “file name does not match module name” issue
This resolves the issue for the ‘jvm-streaming’ package but I could not
reproduce the problem in isolation in our test suite. It makes sense to
assume that the problem should have to do with directory structure/file
names/module names. I re-created all of this in my test and upon inspection
of actual calls to ghc, they were virtually identical (sans concrete module
names). Yet in the case of ‘jvm-streaming’ we get the error message
complaining about the mismatch, while in the other case the test passes.

I further attempted to bring the two cases closer by commenting out ‘deps’
and ‘prebuilt_dependencies’ in the BUILD file of ‘jvm-streaming’, thinking
that since these fields are present in ‘jvm-steraming’ and not in my test,
they may somehow make the difference, but that also did not help.
‘jvm-streaming’ fails and my test passes.

I'm out of ideas, it looks like an inconsistent behavior of GHC.