Fix TemplateHaskell linking against cc_library targets on macOS.
As with the binary, we need to use `install_name_tool` to correct
linker paths so they're relative to the file that references them.

Modifies a test to add a `cc_library` dependency.  Previously all
the tests only depended on Nix-supplied C libraries, which have
absolute paths and thus weren't triggering this bug.