Fix another macOS linking issue.
Previously, a `haskell_test` that dependend transitively on a C library would fail to
run:

```
$ bazel test //tests/library-deps:bin-deps

Abort trap: 6
dyld: Library not loaded:
@executable_path/libtests_Slibrary-deps_Ssublib_Ssublib-c_libsublib-c.so
  Referenced from:
  /private/var/tmp/_bazel_judah.jacobson/82cbe265677c431f236ecc47b43dfa0d/bazel-sandbox/2543326136945263834/execroot/io_tweag_rules_haskell/bazel-out/darwin-fastbuild/bin/tests/library-deps/bin-deps.runfiles/io_tweag_rules_haskell/tests/library-deps/bin-deps
    Reason: image not found
```

Fixed by making the executable linking logic use the same approach as for
Haskell dynamic libraries.  (And splitting the common logic into a helper
function.)

The cause was that the original macOS executable linking logic pointed to the
`solib` symlinks, which are not made available at runtime.  Now, both
executables and libraries point directly at the original solib.