Enable protos from external repositories.
The `short_path` for an external file contains `..` which messes
up the logic around prefixes and also isn't permitted by `protoc`.
Instead, compute it manually from `path`.  Logic emulated from
rules_go:
https://github.com/bazelbuild/rules_go/blob/67f44035d84a352cffb9465159e199066ecb814c/proto/compiler.bzl#L72

This change required marking the test `haskell_toolchain` as `testonly=0`.
The test that I added works on `@com_google_protobuf//:timestamp_proto`, a
`proto_library` which is not testonly; it looks like the aspect that's
generated also can't be testonly.  If this is an issue we can create a new
repo for the test; I just wasn't sure it was worth the effort.