Try to generate Haddocks using just ‘declare_directory’
Here is an attempt to generate Haddock using just ‘declare_directory’. This
however does not work, as nothing is the directory is created for some
reason. We can't also track just Haddock interface file because then we have
a directory and a file, and the directory is a prefix of that file and Bazel
says that both outputs can't be present at the same time.