Split actions.bzl into multiple action files.
The `actions.bzl` file was getting unwieldy (1k+ lines). Furthermore,
this change is preparation for introducing an ancillary `actions`
struct in all `haskell_*` rules, like `ctx.actions`, but that offers
Haskell-specific actions.

Each action should be largely independent and there should be no
dependencies between them. This is not yet the case, but that's
precisely because there are missing abstractions (to be introduced in
future PR's). With the split, it becomes clear which functions are
used by which actions, and which ones are common and therefore need to
be factored out.