Decouple actions from rule context
Actions all took a `ctx` argument. But there was a huge amount of
immutable state hidden in there. This had two problems:

* it means that action implementation is coupled to rule definition,
  when you'd expect actions to be defined largely independently of any
  particular rule. Actions shouldn't have direct access to rule
  attributes, because difficult to mock and test actions in isolation.
* while action implementations are "functional" in the sense of being
  referentially transparent, they implicitly depend on "everything",
  instead of just the things they need being passed as arguments. E.g.
  rule attributes that should only concern linking were passed to
  compilation and vice versa. It was like having functions of type
  `Heap -> Args -> ResultStruct`.

The conjunction of these two problems meant that it was hard to see
what abstractions we should be piecing together. Solution: create
a much smaller derived context and pass that to actions instead. The
derived context, called the "Haskell context", resolves the toolchain,
constructs a standard environment (including `PATH`) to run actions
in, and only contains a very small number of common attributes. Extra
subcontexts are derived for special purposes and passed as explicit
arguments to action functions, e.g. `dep_info`, `cc` and `java`
subcontexts during compilation. So the dependencies of each action is
now more explicit.