Introduce pkg_id struct
In the case of libraries, we unfortunately can't use just the label
for everything. We need a version component, otherwise `ghc-pkg`
complains. We used to have an unholy mishmash of stuffs: package
names, package id's and versions, all disconnected from each other.
Now we package everything into a single struct that we pass along as
an argument.

I inlined `get_library_name` because there's no good place to put that
function, and it's a trivially small one anyways.