Tether haskell toolchain to CC toolchain.
GHC calls GCC itself when it wants to do some linking, assembly, or
preprocessing. The problem is, GHC might not be calling the same GCC
we expect. While we expect the GCC autodected for the Bazel CC
toolchain, GHC uses whatever GCC is found in a `settings` file that it
ships with.

The key is to override the defaults by passing `-pgm*` flags to GHC.
In this way, GHC uses the same GCC as the rest of the Bazel rules,
hence avoiding any potential linker issues.

Fixes #250.