Use the CC toolchain from Nixpkgs rather than the system.
This is necessary for the following commits in this series, because
Nixpkgs GHC assumes a Clang that understands -no-pie. But that's not
always true of the system global one. If whether to pass -no-pie could
be configured on the GHC command-line, we wouldn't have a problem. But
since it's not, we are forced to make sure the CC toolchain matches
what GHC expects. The most reliable way to do that is to draw the CC
toolchain from the exact same place GHC is drawn from: some specific
checkout of Nixpkgs.

Doing this is complicated by the fact that Nixpkgs doesn't currently
provide a Clang that doesn't rely on magic environment variables being
set to work at all. Since Bazel clears the environment in action
sandboxes, we have to reconstruct the environment Nixpkgs Clang
expects. This should no longer be necessary if
https://github.com/NixOS/nixpkgs/issues/42059 is ever resolved.