Reuse Bazel's CC compiler and linker flags.
When calling the CC compiler, or linker, let's do it as closely as
Bazel does it. Because at the end of the day C code should be compiled
consistently, whether done via the native rules or via GHC. But also
because not doing can cause problems. In particular, our cleaning of
the `PATH` makes some GCC's confused about where to find ther linker
and assembler. But Bazel's native rules know how to give GCC a helping
hand. See #259.