Switch to a Nixpkgs that carries at least Bazel v0.14.
We will require >=0.14 in the following commits because that is the
only version that exposes the CC compiler and linker flags in the
toolchain provider. That API will itself change in 0.16 according to
https://github.com/bazelbuild/bazel/issues/4571. Since the new API is
the only way to correctly get the flags we need, we'll probably bump
the minimum required version soon again.