Remove _make_ghc_defs_dump.
Instead of creating our own file to include, we can simply import
includes already in GHC directly. This saves compilation time in the
common case, because we save one additional call to GHC per haskell
library/binary.

Unfortunately, the only complication is that unlike hsc2hs, c2hs does
not automatically add GHC's libdir in the include path. So we have to
do that ourselves.