Make Bazel no longer manage modules
It used to be that we would infer modules from the names of source
files, and then create a dependency graph based on these module names.
But this is,

1. unnecessary and somewhat more complicated,
2. incorrect in the general case, because GHC knows module names
   better than we do.

Both #375 and #168 are cases where module names don't match what we
expect. The solution is to stay out of the business of managing
modules, and manage only packages. Rather than create individual
outputs for each module, we can use `declare_directory` to create
a single output for all the interface files of a library, another for
all the object files etc.