Use ghc-pkg recache instead of ghc-pkg register
Registration of a new package consists in,

1. copying the registration file into the package db,
2. performing some validation on the registration file content,
3. recaching, i.e. regenerating the package db cache file.

Normally, this is all done by `ghc-pkg register`. But in our case,
`ghc-pkg register` is painful, because the validation it performs is
slow, somewhat redundant but especially, too strict (see e.g.
https://ghc.haskell.org/trac/ghc/ticket/15478). So we now do (1) and
(3) manually, by copying then calling `ghc-pkg recache` directly.

The downside is that we do lose the few validations that `ghc-pkg
register` was doing that was useful. e.g. when reexporting modules,
validation checks that the source module does exist. We'll want to
come back to using `ghc-pkg register` eventually, but for now
https://ghc.haskell.org/trac/ghc/ticket/15478 is a blocker.