Drop setting a bogus version number when none needed
In a monorepo use case, version numbers are not useful. All code
always builds against the "latest" version of every component in any
particular snapshot of the monorepo. In fact I'd be up for dropping
the `version` field entirely, if it were not for supporting Bazelified
Cabal packages via Hazel (these need version numbers to correctly
evaluate CPP macros).

Even if version numbers were usually not set, internally we always
used one. But it turns out that GHC is quite happy to handle
versionless packages, so there is no need to set a bogus "1.0.0"
default version number like we did. With this PR, we drop this default
and allow package id's with no version. We don't use version numbers
in package keys. Version numbers only appear in one place: in the
`version` field of a package configuration.

Dropping versions everywhere else is safe to do because all our
packages are already uniquely named across the workspace even without
version numbers.