Fix handling of packages with clashing names
Package names should not need to be unique. After all, that's the very
reason for the package-name / package-id distinction. It turns out
that the root cause for the original problem was very subtle, but very
simple to fix. The problem was that while we were passing
`-hide-all-packages` to GHC during compilation, we were not doing so
during linking. This flag makes sense, and indeed not passing it
causes bizarre `module not found` errors on import. It turns out that
not passing it at link time also causes the bizarre link errors we
were seeing (which only show up when two packages share the same
name).

I can't explain why `-hide-all-packages`, or lack thereof, has this
effect. But I do know that using the same flags during linking as
during compilation is the right thing to do. With this one line change
to the source code, we can continue to have arbitrary non-unique
package names, relying on package id's generated by rules_haskell
internally to be unique.

Closes #395.

Closes #396.