REPL do not depend on their binary/library targets
This fixes the regression related to the issue #289. Several solutions were
proposed they were not satisfactory IMO. I use a different approach here. I
exploit the fact that after loading of source files GHCi knows module names
by itself, so we can use that information to bring them in scope.

How to do that was not entirely obvious though. I've tried several different
approaches and ended up with the one that temporarily redirects stdout to a
file and then calls ‘:show modules’ which prints the list of names of loaded
modules. Then we read the file and extract the names, deleting the file
afterwards.

To load the modules we create a custom GHCi command with ‘:def’ which in
turns makes use of previously obtained module names. After using the command
we undefine it and reload the modules to get rid of any temporary
definitions.