haskell_doc: add 'flags' attribute
This, (rather limited), flag allows the user to pass flags to haddock
to change the haddock behavior, for exemple, to insert unicode symbols.

The implementation is an art of hacking because bazel allows top level
rule to pass arguments to dependency tree of aspect. However the API
only alows string and forces us to provide a known in advance list of
possibles string values.

So for now, limitations are:

- Can only pass a string of flags, separated by space. Formatting
  issues will result in an error
- The list of possible flags in hardcoded (there is two of them, -U
  and --ignore-all-exports). Future update may add more.