Fix GHC toolchain version check
A check whether our actual GHC version coincides with the GHC
version given by the `haskell_toolchain` rule did exist, but just
putting it into `DefaultInfo` did not work as expected.

We move it into `ToolchainInfo` and pass it as input to each GHC
invocation. This way all Haskell targets depend on the version
check (which is only done once, of course, because it only depends on
the toolchain inputs).