Change convention for standing up a REPL.
Where we used to have `foo-repl` implicit targets for all `foo`
targets, we now have use `foo@repl`. The reason is that using `@` is
less likely to cause to cause target name clashes. Furthermore, some
people use snake case while others use kebab case. One advantage of
the `@repl` convention is that it is agnostic to snake vs kebab.

The old convention is kept to avoid making this a breaking change.
`foo-repl` will still work. It's that this is no longer the documented
way to get a REPL.

Drop mention of Bazel v0.12 in documentation. Support for the `@`
character in target names is only available in newer Bazel's anyway.