ci: Limit the number of bazel threads in analysis
By default (until release 0.19), bazel starts as much as 200 threads in
his analysis phase, which means until 200 nix-build calls with
rules_nixpkgs.
This causes a lot of issues, in particular frequent OOMs on the CI and
Nix complaining because it can only build a limited number of
derivations at the same time.