Support linkstatic attribute in libraries
PR #448 introduced the `linkstatic` attribute for `haskell_binary()`.
This change introduces `linkstatic` for `haskell_binary()`. Like for
CC rules, while the attributes have the same name, they have
a different meaning. In the case of libraries, `linkstatic` means
"don't bother creating a shared library".

To support this attribute in libraries, we had to make changes in the
binary code path as well. There are two distinct notions:

* Dynamically linked code.
* Shared libraries (which are dynamically linked).

Binaries can have code built in a special way so that it can call code
in shared Haskell libraries. That is to say, the code is `dynamic` or
not. Library rules can produce (dynamically linked) shared libraries.
That is, we compile `with_shared` or not (both internal variable
names not exposed to the user).