Use the local checkout in start script testing
Previously, the `start` script was always using the latest
rules_haskell release. As it should. But when we're testing it on
a feature branch, the latest rules_haskell version doesn't always
work. If on the branch we update Bazel to a version with breaking
changes, then we need to adapt to those changes in the branch. Which
in turn means the start script should pull in those changes too.