HaskellBuildInfo: restructure external_libraries
In commit 4b5117d1b64aa87476bedb02df078cc20b316d70 external_libraries
was turned to a dict and all list uses correctly converted by using
.values(). It was done in order to support MacOS and the keys were
used in exactly one place. However, it was found out that this wasn’t
working with transitive libraries, so the one use of the keys was
removed again in commit 182df778143d47dbd2dc77a08fb40074fc4e80b5, yet
external_libraries was not converted back to a set. In the following
commits many new usages of external_commits did not take into account
that it is a dict, some treating it as a list, some as a set.

To make the usage more explicit, we convert it to a set of a struct
with two fields, one for the original library and one for the mangled
library.

Fixes https://github.com/tweag/rules_haskell/issues/428