Fix handling of chs modules
The build of `hsndfile` was broken due to an error in the handling of
chs modules. The original chs source file was passed on to the final
`haskell_library` target, instead of the generated hs file.

This change fixes that issue.

The accumulated `module_map` no longer contains a list of all source
files linked or generated in the build tree. In the case of chs files,
the original chs file is linked into the build tree, while the hs file
is generated else where.

Files in the build tree are kept track of in a separate list, so that
the stage that links `extra-source-file` can avoid collisions.