Add failing test-cases for static linking
Static linking fails when depending on a C library indirectly,
due to the order in which the "-l" parameters are passed
to CC when linking. GHC always puts the "-lHSblah" last,
which means if "libHSblah" depends on symbols in another
static library that library needs to come after "-lHSblah"
on the command-line. The proper way to do that is to
set the "extra-libraries" field in the package database.