Add prebuilt-package dependencies to package conf
Previously, static linking (linkstatic = True) of Haskell binary targets
fails if the target has an indirect dependency on a prebuilt package
imported using haskell_import. The reason is that haskell_import has no
access to the package id and the package id of prebuilt package
dependencies is not inserted into the `depends:` section of the package
configuration file of Haskell targets. GHC then does not add linker
flags for the corresponding static libraries and liking fails on missing
symbols. See FormationAI/hazel#64 .

This change fixes that by inserting the package id of prebuilt packages
into the depends: section of Haskell targets.

The corresponding package id is obtained using `ghc-pkg field id` on the
prebuilt package.