Fix //tests/binary-linkstatic-flag for MacOS
`objdump` is not available on the MacOS build. Furthermore, on MacOS
only direct library dependencies are listed under the shared library
dependencies. Also, on MacOS the dynamic library dependency listed in
the binary's header uses the library ID instead of the mangled library
path as on Linux.

This changes and simplifies the binary-linkstatic-flag tests to work on
both Linux and MacOS seemlessly.

First, we turn the indirect C library dependency into a direct
dependency.

Second, instead of checking for shared library dependencies, we directly
check for the `value` and `*HsLib_value_closure` symbols. In a
statically linked binary both those symbols will be defined directly in
the binary. In a dynamically linked binary both symbols will be listed
as undefined. The `nm` tool is available on both the Linux and MacOS
builds and can list all undefined symbols in the binary.