Don't hardcode platforms in toolchain anymore.
Our toolchain implementation was incomplete. We used to assume x86_64
as the architecture, and never set any platform constraints. We now
make it possible to do so in new attributes to `haskell_toolchain`.
This has important implications. It means that we no longer should
define a single toolchain that states different constraints depending
on the host platform. We now instead have multiple toolchains in
scope, one for each operating system. Which properly declared
constraints, the toolchain resolution mechanism can select the right
one.

There is another change in this commit, which we couldn't keep
separate because it was co-dependent to the above change if we want to
stay green. `ghc_bindist` now itself defines a toolchain (no need to
define one separately), and also itself registers it. So `ghc_bindist`
is now easier to use. A side effect of this is that the toolchain
definition for Windows moves into an external dependency, where it
won't bother anyone. If we define the Windows toolchain in `//tests`,
then we'll be fetching and building it even on Linux whenever someone
says `bazel build //tests/...`.

We'll eventually also make Nixpkgs toolchains equally easy to define.