Split out ghc_bindist toolchain definition into own repository
We want the toolchain definition to be tucked away in a separate
repository, that way `bazel build //...` will not match it (and e.g.
build the Windows toolchain even on Linux). At the same time, we don't
want the definition in the bindist repository, because then we need to
download the bindist first before we can see the toolchain definition.
The solution is to add the toolchain definition in its own special
repository.