link.bzl: refactor _add_external_libraries
The interspersed deduplication logic made the intent of the function
unclear, so it is factored into a separate step and an explanatory
comment added.

There is a slight overhead, because `get_lib_name` is called twice per
deduplicated ext_lib, but since `get_lib_name` only does a few string
manipulations that shouldn’t cause problems.

Add a `dedup_on` helper on lists.