Introduce haskell_register_ghc_nixpkgs
This works the same as `haskell_register_ghc_bindists`, except that we
are registering a Nixpkgs package, rather than a bindist. Nixpkgs
toolchains are now gated with the `//haskell/platforms:nixpkgs` config
value. To turn it on, you need to specify the `linux_x86_64_nixpkgs`
or `darwin_x86_64_nixpkgs` platforms.