link.bzl: Rewrite rpath construction
`-rpath` flags passed to GHC are directories that are searched by the
loader (`ld.so` on Linux), to find shared libraries at runtime.

Originally, we used `target_library.path` and went upwards all the way
to the `bazel-out` directory. This fails horribly when sandboxing is
enabled, because only bazel runfiles are visible in that case.
Therefore the first change (“Make rpaths refer to the runfiles
directory”) changed that to only go upwards to the runfiles directory,
using `target_library.short_path` instead of `target_library.path`.

This turned out to not be enough in cases where the dependencies are
in external repositories. In those cases, the `.short_path` contains a
`../` as first element and there is an `external` directory in the
runfiles directory. This commit introduces logic to handle that.

To be honest, this was yet another extremely time-consuming
work-around for the abysmal design of bazel repositories. Most
rulesets will need to implement this logic sooner or later if they use
dynamic linking and want to work correctly.

Fixes //tests/external-haskell-repository.