Test GHC bindist on Linux on CI
We add a new CI job based on a vanilla debian docker image (ie. without
nix in `$PATH`) testing the ghc_bindist Haskell toolchain.

In order to evaluate the `WORKSPACE` file on linux without having
`nix-build` in $PATH, we had to make a couple of changes in the
nix-supported platform detection.

Some tests are depending on some external deps such as hackage, zlib,
doctest, c2hs or the GHC threaded rts. We are ignoring these tests using
a tags filter.