Expose embedded MinGW toolchain in Windows bindists
GHC bindists on Windows include a full C/C++ toolchain, in the
`mingw/` subfolder. GHC uses this toolchain rather than whatever
happens to be in the `PATH`. In fact telling GHC to use whatever is in
the `PATH` will often fail, since there are differences between e.g.
MSYS2 and other installations of MinGW (existence of libmingw32
library etc). The only approach that works is the reverse: tell the
rest of the rules to use GHC's embedded MinGW toolchain.

To do this, we add a simple `CROSSTOOL` file for Windows. On Windows,
run with the following command:

```
$ bazel build //... --crosstool_top=@io_tweag_rules_haskell_ghc_windows_amd64//:toolchain
```