link.bzl: Fix rpaths for haskell_cc_import libraries
`haskell_cc_import` is mostly a no-op that can reference shared
objects from external repositories. The paths of the libraries it adds
follow a different scheme than e.g. cc_* rule dependencies or haskell
dependencies, so we have to account for that in rpath construction
logic.

It provides a `CcSkylarkApiProviderHacked` provider that is added to
its outputs. We can use that to handle libraries imported via
`haskell_cc_import` separately in `create_rpath_entry`.

`haskell_cc_import` should be deprecated by the move to the new C++
skylark API, so we can remove most of this boilerplate in the future.