Don't generate symlinks for source files
Hazel used to generate symbolic links for every source file to place
them into a separate source tree. This was a historical left-over from
when the exact module structure had to be replicated in the file-system.

In effect this created one Bazel target per source file. This was
unnecessarily stressing the distributed cache, as every target is cached
and fetched separately.

This change skips the creation of these symlinks and instead leaves the
source files in their original position in the source tree, and adds
generated files into the source tree in the appropriate place.