Remove --all_incompatible_changes
This comment
https://github.com/bazelbuild/bazel/issues/6861#issuecomment-460145291
and this process:
https://docs.bazel.build/versions/master/backward-compatibility.html
detail the incompatible change process.

`--all_incompatible_changes` is not the right setting for us because
it will:

- enable changes which may never be set by default, hence forcing us
  to do useless work
- enable changes which may be set by default, but at a time in which
  the rest of the bazel ecosystem is not ready yet. See for example
  #639.

Future aware developers of rules_haskell needs to track the bazel
issue tracker for possible incompatible changes and decide if they
justify preliminary work. They may also occasionally enable
`--all_incompatible_changes` to test, but it should not be the default.