linking_utils: realpath symlinked elf dependencies
`ld.so` does something akin to `realpath` on the current elf binary,
and the parent directory of the fully resolved path is then used as
`$ORIGIN`.

`ldd` (which is a bash script around `ld.so`) is actually buggy in
that respect, because it does not `realpath` the file you give it on
the command line, while `ld.so` does (aka the path of the file for
which `ld.so` is used as interpreter).

The `ldd_test` actually broke because of that, with the recent changes
to `RUNPATH` entries.

This means linking_utils is now less buggy than offical GNU/ldd in
that crucial aspect …